refactor(telemetry): nest the event-tracker pillar under its own folder
Move the telemetry wiring into src/configuration/telemetry/event-tracker/ so
upcoming pillars (error-reporter, logger, tracer…) each get their own folder,
mirroring the @arckit/telemetry structure. Consumers import from
'@/configuration/telemetry/event-tracker' directly — no root barrel.

Names follow the pillar: EventTracker (page-view component), EVENT_TRACKER_CONFIG
/ EventTrackerConfig (DI key), alongside trackEvent and the TrackerCategory /
TrackerAction taxonomy.

tsc, biome, 177 tests, next build, vite wc build, depcruise/folderslint green.
