--- src/configuration/telemetry/event-tracker.ts
@@ -1,3 +1,0 @@
-import { matomoBrowserEventTracker } from '@arckit/telemetry/event-tracker';
-
-export const eventTracker = matomoBrowserEventTracker();
