+++ src/configuration/telemetry/event-tracker/event-tracker.tsx
@@ -1,0 +1,1 @@
+export const EventTracker = (): null => null;
