+++ src/configuration/telemetry/event-tracker/index.ts
@@ -1,0 +1,5 @@
+export { EventTracker } from './event-tracker';
+export type { EventTrackerConfig } from './event-tracker-config.key';
+export { EVENT_TRACKER_CONFIG } from './event-tracker-config.key';
+export type { TrackerActionType, TrackerCategoryType } from './track-event';
+export { TrackerAction, TrackerCategory, trackEvent } from './track-event';
