+++ src/configuration/telemetry/event-tracker/register.ts
