--- src/configuration/telemetry/index.ts
@@ -1,5 +1,0 @@
-export type { TrackerActionType, TrackerCategoryType } from './track-event';
-export { TrackerAction, TrackerCategory, trackEvent } from './track-event';
-export { Tracker } from './tracker';
-export type { TrackerConfig } from './tracker-config.key';
-export { TRACKER_CONFIG } from './tracker-config.key';
