--- src/configuration/telemetry/tracker-config.key.ts
@@ -1,8 +1,0 @@
-import { key } from 'piqure';
-
-export type TrackerConfig = {
-  url?: string | undefined;
-  siteId?: string | undefined;
-};
-
-export const TRACKER_CONFIG = key<TrackerConfig>('tracker-config');
