--- src/configuration/telemetry/tracker.tsx
@@ -1,1 +1,0 @@
-export const Tracker = (): null => null;
