+++ src/instrumentation-client.ts
@@ -1,5 +1,5 @@
-import { register } from '@/configuration/telemetry/register';
+import { register } from '@/configuration/telemetry/event-tracker/register';
 
-export { onRouterTransition } from '@/configuration/telemetry/register';
+export { onRouterTransition } from '@/configuration/telemetry/event-tracker/register';
 
 register();
