+++ src/web-components/layout.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from '@arckit/daisyui/theme';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import type { FC, ReactNode } from 'react';
 import { MapProvider } from 'react-map-gl/maplibre';
-import { Tracker } from '@/configuration/telemetry';
+import { EventTracker } from '@/configuration/telemetry/event-tracker';
 import { Navbar } from '@/features/brand/abilities/layout';
 import { Geolocate, SearchAddress } from '@/features/cartographie';
 import {
@@ -22,7 +22,7 @@ const queryClient = new QueryClient({ defaultOptions: { queries: { staleTime: 10
 
 export const WebComponentLayout: FC<WebComponentLayoutProps> = ({ children }) => (
   <QueryClientProvider client={queryClient}>
-    <Tracker />
+    <EventTracker />
     <LoadStreamSync />
     <ThemeProvider attribute='data-theme' defaultTheme='light' enableSystem disableTransitionOnChange>
       <MapProvider>
